+++ src/add/observable/race.ts
@@ -1,0 +1,6 @@
+import {Observable} from '../../Observable';
+import {raceStatic} from '../../operator/race';
+
+Observable.race = raceStatic;
+
+export var _void: void;
