--- src/add/operator/race-static.ts
@@ -1,7 +1,0 @@
-
-import {Observable} from '../../Observable';
-import {race} from '../../operator/race-static';
-
-Observable.race = race;
-
-export var _void: void;
